+++ core/administration.go
@@ -44,6 +44,10 @@ func (app *Application) updateCovid19Config(config *model.COVID19Config) error {
 	return nil
 }
 
+func (app *Application) getAppVersions() ([]string, error) {
+	return app.supportedVersions, nil
+}
+
 func (app *Application) getAllNews() ([]*model.News, error) {
 	news, err := app.storage.ReadNews(0)
 	if err != nil {
@@ -863,6 +867,58 @@ func (app *Application) getRules() ([]*model.Rule, error) {
 		return nil, err
 	}
 	return rules, nil
+}
+
+func (app *Application) getCRules(countyID string, appVersion string) (*model.CRules, error) {
+	if !app.isVersionSupported(appVersion) {
+		return nil, errors.New("app version is not supported")
+	}
+
+	cRules, err := app.storage.FindCRulesByCountyID(appVersion, countyID)
+	if err != nil {
+		return nil, err
+	}
+	return cRules, nil
+}
+
+func (app *Application) updateCRules(current model.User, group string, countyID string, appVersion string, data string) (*model.CRules, error) {
+	cRules, err := app.storage.UpdateCRules(appVersion, countyID, data)
+	if err != nil {
+		return nil, err
+	}
+
+	//audit
+	userIdentifier, userInfo := current.GetLogData()
+	lData := []AuditDataEntry{{Key: "countyID", Value: countyID}, {Key: "appVersion", Value: appVersion}, {Key: "data", Value: data}}
+	defer app.audit.LogUpdateEvent(userIdentifier, userInfo, group, "crules", "", lData)
+
+	return cRules, nil
+}
+
+func (app *Application) getASymptoms(appVersion string) (*model.Symptoms, error) {
+	if !app.isVersionSupported(appVersion) {
+		return nil, errors.New("app version is not supported")
+	}
+
+	symptoms, err := app.storage.ReadSymptoms(appVersion)
+	if err != nil {
+		return nil, err
+	}
+	return symptoms, nil
+}
+
+func (app *Application) updateSymptoms(current model.User, group string, appVersion string, items string) (*model.Symptoms, error) {
+	symptoms, err := app.storage.UpdateSymptoms(appVersion, items)
+	if err != nil {
+		return nil, err
+	}
+
+	//audit
+	userIdentifier, userInfo := current.GetLogData()
+	lData := []AuditDataEntry{{Key: "appVersion", Value: appVersion}, {Key: "items", Value: items}}
+	defer app.audit.LogUpdateEvent(userIdentifier, userInfo, group, "symptoms", "", lData)
+
+	return symptoms, nil
 }
 
 func (app *Application) createRule(current model.User, group string, countyID string, testTypeID string, priority *int,