+++ driven/storage/database.go
@@ -21,6 +21,8 @@ import (
 	"context"
 	"errors"
 	"health/core"
+	"health/core/model"
+	"io/ioutil"
 	"log"
 	"time"
 
@@ -53,8 +55,10 @@ type database struct {
 	counties       *collectionWrapper
 	testtypes      *collectionWrapper
 	rules          *collectionWrapper
-	symptomgroups  *collectionWrapper
-	symptomrules   *collectionWrapper
+	symptomgroups  *collectionWrapper //old
+	symptomrules   *collectionWrapper //old
+	symptoms       *collectionWrapper
+	crules         *collectionWrapper
 	traceexposures *collectionWrapper
 	accessrules    *collectionWrapper
 
@@ -178,6 +182,16 @@ func (m *database) start() error {
 	if err != nil {
 		return err
 	}
+	symptoms := &collectionWrapper{database: m, coll: db.Collection("symptoms")}
+	err = m.applySymptomsChecks(symptoms)
+	if err != nil {
+		return err
+	}
+	crules := &collectionWrapper{database: m, coll: db.Collection("crules")}
+	err = m.applyCRulesChecks(crules, counties)
+	if err != nil {
+		return err
+	}
 	traceexposures := &collectionWrapper{database: m, coll: db.Collection("traceexposures")}
 	err = m.applyTraceExposuresChecks(traceexposures)
 	if err != nil {
@@ -209,6 +223,8 @@ func (m *database) start() error {
 	m.rules = rules
 	m.symptomgroups = symptomgroups
 	m.symptomrules = symptomrules
+	m.symptoms = symptoms
+	m.crules = crules
 	m.traceexposures = traceexposures
 	m.accessrules = accessrules
 
@@ -577,6 +593,95 @@ func (m *database) applySymptomRulesChecks(symptomRules *collectionWrapper) erro
 	}
 
 	log.Println("symptomRules checks passed")
+	return nil
+}
+
+func (m *database) applySymptomsChecks(symptoms *collectionWrapper) error {
+	log.Println("apply symptoms checks.....")
+
+	//add index
+	err := symptoms.AddIndex(bson.D{primitive.E{Key: "app_version", Value: 1}}, false)
+	if err != nil {
+		return err
+	}
+
+	//add initial data for version 2.6 if not added
+	filter := bson.D{primitive.E{Key: "app_version", Value: "2.6"}}
+	var items []*model.Symptom
+	err = symptoms.Find(filter, &items, nil)
+	if err != nil {
+		return err
+	}
+	if len(items) <= 0 {
+		log.Println("there are no symptoms for version 2.6, so we need to add initial data")
+
+		data, err := ioutil.ReadFile("./driven/storage/symptoms_2.6.json")
+		if err != nil {
+			return err
+		}
+		d := model.Symptoms{AppVersion: "2.6", Items: string(data)}
+		_, err = symptoms.InsertOne(&d)
+		if err != nil {
+			return err
+		}
+	} else {
+		log.Println("there are symptoms for version 2.6, so nothing to do")
+	}
+
+	log.Println("symptoms checks passed")
+	return nil
+}
+
+func (m *database) applyCRulesChecks(cRules *collectionWrapper, counties *collectionWrapper) error {
+	log.Println("apply CRules checks.....")
+
+	//add indexes
+	err := cRules.AddIndex(bson.D{primitive.E{Key: "app_version", Value: 1}}, false)
+	if err != nil {
+		return err
+	}
+
+	err = cRules.AddIndex(bson.D{primitive.E{Key: "county_id", Value: 1}}, false)
+	if err != nil {
+		return err
+	}
+
+	//add initial data for version 2.6 and Champaign county if not added
+	//first find the county id
+	chFilter := bson.D{primitive.E{Key: "name", Value: "Champaign"}}
+	var champaignCounty *county
+	err = counties.FindOne(chFilter, &champaignCounty, nil)
+	if err != nil {
+		return err
+	}
+	if champaignCounty == nil {
+		return errors.New("there is no a Champaign county")
+	}
+
+	//check if added
+	filter := bson.D{primitive.E{Key: "app_version", Value: "2.6"}, primitive.E{Key: "county_id", Value: champaignCounty.ID}}
+	var items []*model.CRules
+	err = cRules.Find(filter, &items, nil)
+	if err != nil {
+		return err
+	}
+	if len(items) <= 0 {
+		log.Println("there are no symptoms rules for version 2.6 and Champaign county, so we need to add initial data")
+
+		data, err := ioutil.ReadFile("./driven/storage/rules_2.6.json")
+		if err != nil {
+			return err
+		}
+		d := model.CRules{AppVersion: "2.6", CountyID: champaignCounty.ID, Data: string(data)}
+		_, err = cRules.InsertOne(&d)
+		if err != nil {
+			return err
+		}
+	} else {
+		log.Println("there are symptoms rules for version 2.6 and Champaign county, so nothing to do")
+	}
+
+	log.Println("CRules checks passed")
 	return nil
 }