+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Expose Rules APIs. [#28](https://github.com/rokwire/health-building-block/issues/28)
 
 ## [1.15.0] - 2020-09-10
 ### Added